Rename files and fix error
