+++ chistyakov.vladislav/P1/sequencenumofmax.hpp
